Extract `filterUnsafeFilenameChars` and `keepFileExtensionAndFilter`
