+++ helpers/local/filterUnsafeFilenameChars.js
@@ -1,0 +1,15 @@
+const filterUnsafeFilenameChars = input => {
+  if (!input) {
+    return '';
+  }
+
+  const text = input
+    .replace(/[=&<>:'"/\\|?*]/g, ' ')
+    .replace(/\s+/g, '-')
+    .slice(0, 245);
+  return text;
+};
+
+module.exports = {
+  filterUnsafeFilenameChars,
+};
